server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
google-cloud-aiplatform==1.47.0
google-generativeai==0.8.3
yagmail==0.15.293
PyJWT==2.8.0
google-auth>=2.20.0